+++ devel/dqDevel.R
@@ -1,8 +1,9 @@
 ncOrig <- '../Data/ncmod/original.nc'
 ncMod <- '../Data/ncmod/modified.nc'
 
+ncOrig <- '../Data/ncmod/NEFSC_GOM_202510_USTR09_20251026.nc'
 library(ncdf4)
-nc <- nc_open(ncMod, write = TRUE)
+nc <- nc_open(ncOrig, write = TRUE)
 # ok lets try making the DQ mat
 # should be vars$quality_flag
 inSize <- nc$var$psd$size
@@ -19,28 +20,7 @@ library(PAMscapes)
 modData <- loadSoundscapeData(ncMod, keepQuals = c(1,2))
 hm <- nc_open('tutorial/MANTAExample.nc')
 
-addDataQuality <- function(ncFile, freqRange=NULL, timeRange=NULL, value=2, verbose=FALSE) {
-    nc <- nc_open(ncFile)
-    on.exit(nc_close(nc))
-    if(!'quality_flag' %in% names(nc$var)) {
-        if(verbose) {
-            cat('No "quality_flag" variable found, creating a new one')
-        }
-        inSize <- nc$var$psd$size
-        dqVar <- ncvar_def('quality_flag', prec='byte', longname='Data quality flag', dim=nc$var$psd$dim, units='')
-        nc <- ncvar_add(nc, dqVar)
-        dqMat <- matrix(2, nrow=inSize[1], ncol=inSize[2])
-    } else {
-        dqMat <- ncvar_get(nc, 'quality_flag', start=c(1, 1), count=c(-1, -1))
-    }
-    times <- PAMscapes:::ncTimeToPosix(nc$dim$time)
-    freqs <- nc$dim$frequency$vals
-    type <- PAMscapes:::checkFreqType(freqs)
-    dqMat <- markDQMatrix(dqMat, freqRange=freqRange, timeRange=timeRange, value=value,
-                          times=times, freqs=freqs)
 
-    ncvar_put(nc, varid='quality_flag', vals=dqMat)
-}
 
 
 
@@ -71,7 +51,7 @@ dqdf <- cbind(dq$time, dqdf)
 names(dqdf)[1] <- 'UTC'
 dqLong <- longQuality(dqdf)
 
-g +
+g <- ggplot() +
     geom_rect(data=dqLong, aes(xmin=UTC,
                                xmax=UTCend,
                                ymin=freqLow,
@@ -96,4 +76,38 @@ ggplot() +
 library(patchwork)
 g / dq
 
-# okaaaaay dq plotters devel
+## we back in 2026 ----
+ncOrig <- '../Data/DQ/NEFSC_GOM_202309_USTR01_20231029.nc'
+data <- loadSoundscapeData(ncOrig)
+annoFile <- '../Data/DQ/DailyAnnotations.csv'
+anno <- readr::read_csv(annoFile)
+intervals <- readr::read_csv('../Data/DQ/makara_intervals.csv') %>% 
+    filter(deployment_code == 'NEFSC_GOM_202309_USTR01')
+qs <- addDataQuality(ncOrig, annotation = intervals, reset=TRUE, test=F, updateUnknown = 1)
+table(qs)
+image(t(qs), useRaster = TRUE)
+loadSoundscapeData(ncOrig, keepQuals = c(1)) %>% 
+    plotLTSA(bin='30min')
+
+nc <- nc_open(ncOrig)
+nc
+nc_close(nc)
+# need to go directly from annotainos -> update NC file
+# could be  button in the app, could happen later (if later, may need way to update paths)
+# Need to make Makara -> Annotations, but do we also need Annotations -> Makara?
+# Need to add some sort of metadata comment like "we messed with thsi file in PAMscapes"
+# Add a plot quality check box to the app 
+# Update all 2 flags to some other value
+# More general global att editor/updater side of tool not just for this
+
+ncDir <- '../Data/DQ'
+library(PAMscapes)
+anno <- readr::read_csv('../Data/DQ/DailyAnnotations.csv')
+intervals <- readr::read_csv('../Data/DQ/makara_intervals.csv')
+runDailyLTSAReview(ncDir)
+
+# ncatt_get(nc, 0) lists all global atts
+# ncatt_put(nc, 0, 'name', value)
+
+# markDQ expects 3 lists of equal length (or 1, to repeat)
+